requirements.txt
requests
pdfplumber
openai
python-dotenv
supabase
urllib3
